@@ -1,28 +1,25 @@
-'use strict';
-
-var React = require('react');
-var {
-  PropTypes,
-} = React;
-var ReactNative = require('react-native');
-var {
+import React, { PropTypes } from 'react';
+import {
   EdgeInsetsPropType,
   Platform,
-  ReactNativeViewAttributes,
   View,
   Animated,
   requireNativeComponent,
   NativeModules,
   ColorPropType,
-} = ReactNative;
+  findNodeHandle,
+} from 'react-native';
+var NativeMethodsMixin = require('react/lib/NativeMethodsMixin');
 
-var MapMarker = require('./MapMarker');
-var MapPolyline = require('./MapPolyline');
-var MapPolygon = require('./MapPolygon');
-var MapCircle = require('./MapCircle');
-var MapCallout = require('./MapCallout');
+import MapMarker from './MapMarker';
+import MapPolyline from './MapPolyline';
+import MapPolygon from './MapPolygon';
+import MapCircle from './MapCircle';
+import MapCallout from './MapCallout';
 
-var MapView = React.createClass({
+// eslint-disable-next-line react/prefer-es6-class
+const MapView = React.createClass({
+  mixins: [NativeMethodsMixin],
 
   viewConfig: {
     uiViewClassName: 'AIRMap',
@@ -320,13 +317,13 @@ var MapView = React.createClass({
 
   },
 
-  getInitialState: function() {
+  getInitialState() {
     return {
       isReady: Platform.OS === 'ios',
     };
   },
 
-  componentDidMount: function() {
+  componentDidMount() {
     const { region, initialRegion } = this.props;
     if (region && this.state.isReady) {
       this.refs.map.setNativeProps({ region });
@@ -335,9 +332,9 @@ var MapView = React.createClass({
     }
   },
 
-  componentWillUpdate: function(nextProps) {
-    var a = this.__lastRegion;
-    var b = nextProps.region;
+  componentWillUpdate(nextProps) {
+    const a = this.__lastRegion;
+    const b = nextProps.region;
     if (!a || !b) return;
     if (
       a.latitude !== b.latitude ||
@@ -349,7 +346,7 @@ var MapView = React.createClass({
     }
   },
 
-  _onMapReady: function() {
+  _onMapReady() {
     const { region, initialRegion } = this.props;
     if (region) {
       this.refs.map.setNativeProps({ region });
@@ -359,7 +356,7 @@ var MapView = React.createClass({
     this.setState({ isReady: true });
   },
 
-  _onLayout: function(e) {
+  _onLayout(e) {
     const { region, initialRegion, onLayout } = this.props;
     const { isReady } = this.state;
     if (region && isReady && !this.__layoutCalled) {
@@ -372,7 +369,7 @@ var MapView = React.createClass({
     onLayout && onLayout(e);
   },
 
-  _onChange: function(event: Event) {
+  _onChange(event: Event) {
     this.__lastRegion = event.nativeEvent.region;
     if (event.nativeEvent.continuous) {
       this.props.onRegionChange &&
@@ -383,34 +380,34 @@ var MapView = React.createClass({
     }
   },
 
-  animateToRegion: function (region, duration) {
+  animateToRegion(region, duration) {
     this._runCommand('animateToRegion', [region, duration || 500]);
   },
 
-  animateToCoordinate: function (latLng, duration) {
+  animateToCoordinate(latLng, duration) {
     this._runCommand('animateToCoordinate', [latLng, duration || 500]);
   },
 
-  fitToElements: function(animated) {
+  fitToElements(animated) {
     this._runCommand('fitToElements', [animated]);
   },
 
-  fitToSuppliedMarkers: function(markers, animated) {
+  fitToSuppliedMarkers(markers, animated) {
     this._runCommand('fitToSuppliedMarkers', [markers, animated]);
   },
 
-  takeSnapshot: function (width, height, region, callback) {
+  takeSnapshot(width, height, region, callback) {
     if (!region) {
       region = this.props.region || this.props.initialRegion;
     }
     this._runCommand('takeSnapshot', [width, height, region, callback]);
   },
 
-  _getHandle: function() {
-    return ReactNative.findNodeHandle(this.refs.map);
+  _getHandle() {
+    return findNodeHandle(this.refs.map);
   },
 
-  _runCommand: function (name, args) {
+  _runCommand(name, args) {
     switch (Platform.OS) {
       case 'android':
         NativeModules.UIManager.dispatchViewManagerCommand(
@@ -429,7 +426,7 @@ var MapView = React.createClass({
     }
   },
 
-  render: function() {
+  render() {
     let props;
 
     if (this.state.isReady) {
@@ -462,7 +459,7 @@ var MapView = React.createClass({
   },
 });
 
-var AIRMap = requireNativeComponent('AIRMap', MapView, {
+let AIRMap = requireNativeComponent('AIRMap', MapView, {
   nativeOnly: {
     onChange: true,
     onMapReady: true,
