@@ -1,6 +1,6 @@
-var React = require('react');
-var ReactNative = require('react-native');
-var {
+let React = require('react');
+const ReactNative = require('react-native');
+let {
   StyleSheet,
   PropTypes,
   View,
@@ -11,9 +11,9 @@ var {
   Platform,
 } = ReactNative;
 
-var MapView = require('react-native-maps');
+let MapView = require('react-native-maps');
 
-var screen = Dimensions.get('window');
+const screen = Dimensions.get('window');
 
 const ASPECT_RATIO = screen.width / screen.height;
 const LATITUDE = 37.78825;
@@ -21,7 +21,7 @@ const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 
-var AnimatedMarkers = React.createClass({
+const AnimatedMarkers = React.createClass({
   getInitialState() {
     return {
       coordinate: new Animated.Region({
@@ -32,7 +32,7 @@ var AnimatedMarkers = React.createClass({
   },
 
   animate() {
-    var { coordinate } = this.state;
+    const { coordinate } = this.state;
     coordinate.timing({
       latitude: LATITUDE + (Math.random() - 0.5) * LATITUDE_DELTA / 2,
       longitude: LONGITUDE + (Math.random() - 0.5) * LONGITUDE_DELTA / 2,
@@ -66,8 +66,7 @@ var AnimatedMarkers = React.createClass({
 });
 
 
-
-var styles = StyleSheet.create({
+let styles = StyleSheet.create({
   container: {
     position: 'absolute',
     top: 0,
