@@ -1,10 +1,10 @@
-var React = require('react');
-var ReactNative = require('react-native');
-var {
+let React = require('react');
+const ReactNative = require('react-native');
+const {
   PropTypes,
 } = React;
 
-var {
+let {
   StyleSheet,
   View,
   Text,
@@ -12,13 +12,12 @@ var {
 } = ReactNative;
 
 const PriceMarker = ({ amount, selected, style }) => {
-
-  var background = selected.interpolate({
+  const background = selected.interpolate({
     inputRange: [0, 1],
     outputRange: ['#FF5A5F', '#4da2ab'],
   });
 
-  var border = selected.interpolate({
+  const border = selected.interpolate({
     inputRange: [0, 1],
     outputRange: ['#D23F44', '#007a87'],
   });
@@ -26,19 +25,18 @@ const PriceMarker = ({ amount, selected, style }) => {
   return (
     <Animated.View style={[styles.container, style]}>
       <Animated.View style={[styles.bubble, {
-          backgroundColor: background,
-          borderColor: border,
-        }]}
-        >
+        backgroundColor: background,
+        borderColor: border,
+      }]}>
         <Text style={styles.dollar}>$</Text>
         <Text style={styles.amount}>{amount}</Text>
       </Animated.View>
       <Animated.View style={[styles.arrowBorder, {
-          borderTopColor: border,
-        }]} />
+        borderTopColor: border,
+      }]} />
       <Animated.View style={[styles.arrow, {
-          borderTopColor: background,
-        }]} />
+        borderTopColor: background,
+      }]} />
     </Animated.View>
   );
 };
