@@ -1,6 +1,6 @@
-var React = require('react');
-var ReactNative = require('react-native');
-var {
+let React = require('react');
+const ReactNative = require('react-native');
+let {
   StyleSheet,
   PropTypes,
   View,
@@ -9,9 +9,9 @@ var {
   TouchableOpacity,
 } = ReactNative;
 
-var MapView = require('react-native-maps');
+let MapView = require('react-native-maps');
 
-var { width, height } = Dimensions.get('window');
+let { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
@@ -19,7 +19,7 @@ const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 
-var DisplayLatLng = React.createClass({
+const DisplayLatLng = React.createClass({
   getInitialState() {
     return {
       region: {
@@ -44,7 +44,7 @@ var DisplayLatLng = React.createClass({
   },
 
   randomRegion() {
-    var { region } = this.state;
+    const { region } = this.state;
     return {
       ...this.state.region,
       latitude: region.latitude + (Math.random() - 0.5) * region.latitudeDelta / 2,
@@ -64,7 +64,7 @@ var DisplayLatLng = React.createClass({
         >
         </MapView>
         <View style={[styles.bubble, styles.latlng]}>
-          <Text style={{ textAlign: 'center'}}>
+          <Text style={{ textAlign: 'center' }}>
             {`${this.state.region.latitude.toPrecision(7)}, ${this.state.region.longitude.toPrecision(7)}`}
           </Text>
         </View>
@@ -81,7 +81,7 @@ var DisplayLatLng = React.createClass({
   },
 });
 
-var styles = StyleSheet.create({
+let styles = StyleSheet.create({
   container: {
     position: 'absolute',
     top: 0,
