@@ -1,6 +1,6 @@
-var React = require('react');
-var ReactNative = require('react-native');
-var {
+let React = require('react');
+const ReactNative = require('react-native');
+let {
   StyleSheet,
   PropTypes,
   View,
@@ -10,24 +10,24 @@ var {
   ScrollView,
 } = ReactNative;
 
-var MapView = require('react-native-maps');
-var PriceMarker = require('./PriceMarker');
+let MapView = require('react-native-maps');
+let PriceMarker = require('./PriceMarker');
 
-var { width, height } = Dimensions.get('window');
+let { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
 const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
-var id = 0;
+let id = 0;
 
-var Event = React.createClass({
+let Event = React.createClass({
   shouldComponentUpdate(nextProps) {
     return this.props.event.id !== nextProps.event.id;
   },
   render() {
-    var { event } = this.props;
+    const { event } = this.props;
     return (
       <View style={styles.event}>
         <Text style={styles.eventName}>{event.name}</Text>
@@ -37,7 +37,7 @@ var Event = React.createClass({
   },
 });
 
-var DisplayLatLng = React.createClass({
+const DisplayLatLng = React.createClass({
   getInitialState() {
     return {
       region: {
@@ -53,14 +53,14 @@ var DisplayLatLng = React.createClass({
   makeEvent(e, name) {
     return {
       id: id++,
-      name: name,
+      name,
       data: e.nativeEvent ? e.nativeEvent : e,
     };
   },
 
   recordEvent(name) {
     return e => {
-      var { events } = this.state;
+      const { events } = this.state;
       this.setState({
         events: [
           this.makeEvent(e, name),
@@ -115,7 +115,7 @@ var DisplayLatLng = React.createClass({
   },
 });
 
-var styles = StyleSheet.create({
+let styles = StyleSheet.create({
   container: {
     position: 'absolute',
     top: 0,
