@@ -1,6 +1,6 @@
-var React = require('react');
-var ReactNative = require('react-native');
-var {
+let React = require('react');
+const ReactNative = require('react-native');
+let {
   StyleSheet,
   PropTypes,
   View,
@@ -10,10 +10,10 @@ var {
   Image,
 } = ReactNative;
 
-var MapView = require('react-native-maps');
-var PriceMarker = require('./PriceMarker');
+let MapView = require('react-native-maps');
+const PriceMarker = require('./PriceMarker');
 
-var { width, height } = Dimensions.get('window');
+let { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
@@ -22,7 +22,7 @@ const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 const SPACE = 0.01;
 
-var MarkerTypes = React.createClass({
+const MarkerTypes = React.createClass({
   render() {
     return (
       <View style={styles.container}>
@@ -35,7 +35,7 @@ var MarkerTypes = React.createClass({
             latitudeDelta: LATITUDE_DELTA,
             longitudeDelta: LONGITUDE_DELTA,
           }}
-          >
+        >
           <MapView.Marker
             coordinate={{
               latitude: LATITUDE + SPACE,
@@ -55,14 +55,14 @@ var MarkerTypes = React.createClass({
             centerOffset={{ x: -42, y: -60 }}
             anchor={{ x: 0.84, y: 1 }}
             image={require('./assets/flag-pink.png')}
-            />
+          />
         </MapView>
       </View>
     );
   },
 });
 
-var styles = StyleSheet.create({
+let styles = StyleSheet.create({
   container: {
     position: 'absolute',
     top: 0,
@@ -83,7 +83,7 @@ var styles = StyleSheet.create({
     marginLeft: 33,
     marginTop: 18,
     fontWeight: 'bold',
-  }
+  },
 });
 
 module.exports = MarkerTypes;
