@@ -1,6 +1,6 @@
-var React = require('react');
-var ReactNative = require('react-native');
-var {
+let React = require('react');
+const ReactNative = require('react-native');
+let {
   StyleSheet,
   PropTypes,
   View,
@@ -9,19 +9,19 @@ var {
   TouchableOpacity,
 } = ReactNative;
 
-var MapView = require('react-native-maps');
-var PriceMarker = require('./PriceMarker');
+let MapView = require('react-native-maps');
+const PriceMarker = require('./PriceMarker');
 
-var { width, height } = Dimensions.get('window');
+let { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
 const LATITUDE = 37.78825;
 const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
-var id = 0;
+let id = 0;
 
-var PolylineCreator = React.createClass({
+const PolylineCreator = React.createClass({
   getInitialState() {
     return {
       region: {
@@ -36,7 +36,7 @@ var PolylineCreator = React.createClass({
   },
 
   finish() {
-    var { polygons, editing } = this.state;
+    let { polygons, editing } = this.state;
     this.setState({
       polygons: [...polygons, editing],
       editing: null,
@@ -44,7 +44,7 @@ var PolylineCreator = React.createClass({
   },
 
   onPanDrag(e) {
-    var { editing } = this.state;
+    const { editing } = this.state;
     if (!editing) {
       this.setState({
         editing: {
@@ -105,7 +105,7 @@ var PolylineCreator = React.createClass({
   },
 });
 
-var styles = StyleSheet.create({
+let styles = StyleSheet.create({
   container: {
     position: 'absolute',
     top: 0,
