@@ -1,3 +1,3 @@
-var MapView = require('./components/MapView');
+const MapView = require('./components/MapView');
 
 module.exports = MapView;
